sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adp-sdk"
version = "0.0.1"
description = "ADP SDK (Python)"
readme = "README.md"
authors = [{name = "ACME"}]
requires-python = ">=3.11"
dependencies = ["pydantic>=2.6", "pyyaml>=6.0", "jsonschema>=4.21"]

[tool.setuptools]
packages = ["adp_sdk"]

[project.optional-dependencies]
test = ["pytest"]
